pyproject.toml
[tool.poetry]
name = "tangram"
version = "0.1.0"
description = ""
authors = [
  "Xavier Olive <git@xoolive.org>",
  "Junzi Sun <junzi.sun@gmail.com>",
  "Michel Khalaf <khalafmichel98@gmail.com>",
  "Xiaogang Huang <maple.hl@gmail.com>",
]
license = "MIT"

[tool.poetry.scripts]
tangram = 'tangram.__main__:main'

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
appdirs = "^1.4.4"
fastapi = "^0.110.1"
uvicorn = { extras = ["standard"], version = "^0.30.1" }
broadcaster = "^0.3.1"
websockets = "^12.0"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.1"
httpx = "^0.27.0"
pandas = "^2.2.2"
jinja2 = "^3.1.4"
eval-type-backport = { version = "^0.2.0", python = "<=3.9" }
pydantic-settings = "^2.2.1"
anyio = { extras = ["trio"], version = "^4.4.0" }
redis = "^5.0.8"

[tool.poetry.group.dev.dependencies]
mypy = ">=0.950"
ruff = "^0.4.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.11"
platform = "posix"

color_output = true
pretty = true
show_column_numbers = true
strict = true

# probably redundant, but useful when switching strict to false
check_untyped_defs = true
ignore_missing_imports = true
warn_no_return = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unused_ignores = true

[tool.ruff]
exclude = [
  '.git',
  '.direnv',
  '.venv',
  'venv',
  '.mypy_cache',
  '.pyenv',
  '.pytype',
  '.vscode',
  '__pypackages__',
  '.dist',
  'result',
  'site-packages',
  'node_modules',
]

line-length = 120
indent-width = 4

target-version = 'py311'

[tool.ruff.lint]
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"
